+++ src/components/app/app.tsx
@@ -1,7 +1,7 @@
 import {Route, Routes} from 'react-router-dom';
 import { AppRoute, AuthorizationStatus } from '../../const';
 import WelcomeScreen from '../../pages/welcome-screen/welcome-screen';
-import NotFound from '../../pages/not-found-screen/not-found-screen';
+import NotFoundScreen from '../../pages/not-found-screen/not-found-screen';
 import FavoritesScreen from '../../pages/favorites-screen/favorites-screen';
 import LoginScreen from '../../pages/login-screen/login-screen';
 import OfferScreen from '../../pages/offer-screen/offer-screen';
@@ -12,15 +12,18 @@ import { useAppSelector } from '../../hooks/index';
 import LoadingScreen from '../../pages/loading-screen/loading-screen';
 import HistoryRouter from '../history-route/history-route';
 import browserHistory from '../../browser-history';
+import AuthorizationNav from '../authorization-header/authorization-header';
+import { TOffer } from '../../types/offer';
 
 
 type AppScreenProps = {
-  // offers: TOffer[];
+  offersType: TOffer[];
   reviews: TReview[];
   cities: string[];
 }
 
-function App({ reviews, cities}: AppScreenProps): JSX.Element {
+function App({ cities}: AppScreenProps): JSX.Element {
+
   const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
   const isOffersDataLoading = useAppSelector((state) => state.isOffersDataLoading);
   const offers = useAppSelector((state) => state.offers);
@@ -31,20 +34,35 @@ function App({ reviews, cities}: AppScreenProps): JSX.Element {
     );
   }
 
-
   return (
     <HelmetProvider>
-      <HistoryRouter history={browserHistory}>
-        <Routes>
-          <Route path={AppRoute.Main} element={<WelcomeScreen offers = {offers} cities={cities} />} />
-          <Route path={AppRoute.Login} element={<LoginScreen />} />
-          <Route path={AppRoute.Favotites} element={<PrivateRoute authorizationStatus={AuthorizationStatus.Auth}><FavoritesScreen offers={offers} /></PrivateRoute>} />
-          <Route path={`${AppRoute.Offer}:id`}element={<OfferScreen offers={offers} reviews={reviews} />} />
-          <Route path="*" element={<NotFound />} />
-        </Routes>
-      </HistoryRouter>
-    </HelmetProvider>
-  );
+    <HistoryRouter history={browserHistory}>
+      <Routes>
+        <Route path={AppRoute.Main} element={<AuthorizationNav authorizationStatus={authorizationStatus} />}>
+          <Route
+            index element={
+              <WelcomeScreen
+                offers={offers}
+                cities={cities}
+              />
+            }
+          />
+          <Route element={<PrivateRoute authorizationStatus={authorizationStatus} />}>
+            <Route
+              element={<FavoritesScreen />}
+              path={AppRoute.Favotites}
+            />
+          </Route>
+          <Route
+            path={`${AppRoute.Offer}:id`}
+            element={<OfferScreen />}
+          />
+          <Route path="*" element={<NotFoundScreen />} />
+        </Route>
+        <Route path={AppRoute.Login} element={<LoginScreen />} />
+      </Routes>
+    </HistoryRouter>
+  </HelmetProvider>
+);
 }
-
 export default App;
