+++ src/components/private-route/private-route.tsx
@@ -1,19 +1,10 @@
-import { Navigate } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+import { Navigate, Outlet } from 'react-router-dom';
+import { AuthorizationStatus } from '../../const';
 
 type PrivateRouteProps = {
   authorizationStatus: AuthorizationStatus;
-  children: JSX.Element;
 }
 
-function PrivateRoute(props: PrivateRouteProps): JSX.Element {
-  const {authorizationStatus, children} = props;
-
-  return (
-    authorizationStatus === AuthorizationStatus.Auth
-      ? children
-      : <Navigate to={AppRoute.Login} />
-  );
-}
+const PrivateRoute = ({authorizationStatus}: PrivateRouteProps) => authorizationStatus === AuthorizationStatus.Auth ? <Outlet /> : <Navigate to="/login" />;
 
 export default PrivateRoute;
